Added argparse to openset algos
