+++ opensetAlgos/__init__.py
@@ -1,6 +1,6 @@
-from .openmax import OpenMax_Training
-from .openmax import OpenMax_Inference
-from .openmax import MultiModalOpenMax_Training
-from .openmax import MultiModalOpenMax_Inference
-from .EVM import EVM_Training
-from .EVM import EVM_Inference
+__all__=['OpenMax_Params','OpenMax_Training','OpenMax_Inference',
+         'MultiModalOpenMax_Params','MultiModalOpenMax_Training','MultiModalOpenMax_Inference',
+         'EVM_Params','EVM_Training','EVM_Inference']
+from .openmax import *
+from .EVM import *
+from .multimodal_openmax import *
